requirements.txt
pyserial==3.5
numpy==1.26.2
opencv-python==4.8.1.78
python-ev3dev2==2.1.0.post1
Pillow==10.1.0